feat(bin/server): Do not allow to create first user with that tool
When you create the first user with this tool, this user is not an admin user and does not have permission to unlock the ubud keychain. However, this means that the keychain can never be used, which means that plugins will not work, for example. The first user is correctly created when the server is started, and the e-mail address and password are output in the logs.